+++ src/api/src/controllers/PostController.ts
@@ -24,3 +24,35 @@ export const getPostById = async (req: Request, res: Response) => {
         processError(error, res);
     }
 }
+
+export const createPost = async (req: Request, res: Response) => {
+   try {
+        const result = await PostService.createPost(req.body);
+
+        res.send(result)
+    } catch (error) {
+       processError(error, res);
+    }
+}
+
+export const deletePost = async (req: Request, res: Response) => {
+   try {
+        const id = Number(req.params.id);
+        const result = await PostService.deletePost(id);
+
+        res.send(result)
+    } catch (error) {
+       processError(error, res);
+    }
+}
+
+export const updatePost = async (req: Request, res: Response) => {
+   try {
+        const id = Number(req.params.id);
+        const result = await PostService.updatePost(id, req.body);
+
+        res.send(result)
+    } catch (error) {
+       processError(error, res);
+    }
+}