+++ src/api/src/repository/PostRepository.ts
@@ -3,6 +3,8 @@ import ExtError from '../util/errors/ExtError';
 
 import HTTP_STATUS from '../enum/HttpStatus';
 
+import { PostType } from '../types/post';
+
 const prisma = new PrismaClient();
 const postsTable = prisma.post;
 
@@ -22,6 +24,56 @@ export const getPostById = async (id: number) => {
     if(!post) {
         throw new ExtError(HTTP_STATUS.NOT_FOUND, "Post with the given ID was not found.")
     }
+
+    return post;
+}
+
+export const createPost = async ({ title, excerpt, img, author_id }: PostType) => {
+    const post = await postsTable.create({
+        data: {
+            title,
+            excerpt,
+            img,
+            author_id
+        }
+    })
+
+    if(!post) {
+        throw new ExtError(HTTP_STATUS.BAD_REQUEST, "An error occured, post body not valid.")
+    }
+
+    return post;
+}
+
+export const deletePost = async (id: number) => {
+    const postExists = await getPostById(id);
+
+    if(!postExists) {
+        throw new ExtError(HTTP_STATUS.NOT_FOUND, "Post with the given ID was not found.")
+    }
+
+    const post = await postsTable.delete({
+        where: {
+            id
+        }
+    })
+
+    return post;
+}
+
+export const updatePost = async (id: number, data: PostType) => {
+    const postExists = await getPostById(id);
+
+    if(!postExists) {
+        throw new ExtError(HTTP_STATUS.NOT_FOUND, "Post with the given ID was not found.")
+    }
+
+    const post = await postsTable.update({
+        where: {
+            id
+        },
+        data
+    })
 
     return post;
 }