+++ src/api/src/services/UserService.ts
@@ -1,8 +1,7 @@
-import HTTP_STATUS from "../enum/HttpStatus";
 import * as userRepository from "../repository/UserRepository";
+import { UserType } from '../types/user';
 
 import { generatePasswordHash } from "../util/authentication/authenticationFunctions";
-import ExtError from "../util/errors/ExtError";
 
 export const getAllUsers = async () => {
     const users = await userRepository.getAllUsers();
@@ -10,31 +9,25 @@ export const getAllUsers = async () => {
     return users;
 }
 
-export const getUserById = async (id: string) => {
-    const idAsNumber = parseInt(id);
-    const user = await userRepository.getUserById(idAsNumber);
+export const getUserById = async (id: number) => {
+    const user = await userRepository.getUserById(id);
 
     return user;
 }
 
-export const doesUserWithGivenEmailExists = async (email: string) => {
-    const user = await userRepository.getUserByEmail(email);
-
-    if(user)
-        return true;
-    
-    return false;
-}
-
-export const createUser = async (data: any) => {
-    const { email, password } = data;
-    
-    if(await doesUserWithGivenEmailExists(email))
-        throw new ExtError(HTTP_STATUS.BAD_REQUEST, "The user with the given email already exists");
-
+export const createUser = async (data: UserType) => {
+    const { password } = data;
+ 
     const password_hash = generatePasswordHash(password);
     data.password = password_hash;
 
     return await userRepository.createUser(data);
 }
 
+export const deleteUser = async (id: number) => {
+    return await userRepository.deleteUser(id);
+}
+
+export const updateUser = async (id: number, data: UserType) => {
+    return await userRepository.updateUser(id, data);
+}