+++ src/client/src/App.tsx
@@ -1,9 +1,11 @@
 import { Routes, Route } from 'react-router-dom';
 
 import Header from './components/Header';
-import Blog from './components/Blog';
+
+import Blog from './pages/Blog/Blog';
+import Dashboard from './pages/Dashboard/Dashboard';
 import Login from './pages/Login/Login';
-import Signup from './components/Signup';
+import Signup from './pages/Signup/Signup';
 
 function App() {
 
@@ -14,6 +16,7 @@ function App() {
 				<Route path='/' element={<Blog />} />
 				<Route path='/login' element={<Login />} />
 				<Route path='/signup' element={<Signup />} />
+				<Route path='/dashboard' element={<Dashboard />} />
 			</Routes>
 		</>
 	);