+++ src/client/src/pages/Blog/BlogPost.tsx
@@ -1,14 +1,8 @@
-type BlogPostProps = {
-	item: {
-		title: string,
-		excerpt: string,
-		img: string
-	}
-}
+import { BlogPostProps } from '../../types/PropTypes';
 
 function BlogPost({item: {title, excerpt, img}}: BlogPostProps) {
 	return (
-		<article className='max-w-md mx-auto mt-4 shadow-md rounded-md duration-300 hover:shadow-xl cursor-pointer'>
+		<article className='min-w-full max-w-md mx-auto mt-4 shadow-md rounded-md duration-300 hover:shadow-xl cursor-pointer'>
 			<a href='/'>
 				<img
 					src={img}