+++ src/client/src/types/PropTypes.ts
@@ -1,5 +1,13 @@
 import { ReactElement } from 'react';
 
+export type BlogPostProps = {
+	item: {
+		title: string,
+		excerpt: string,
+		img: string
+	}
+}
+
 export type InputProps = {
 	id: string,
 	type: string,